feat: connected game engine to front-end
This is a deceptively simple statement.
Some serious thinking went into how to do it.
First I had to realize that the browser is an event driven.
The engine is a step by step and the game controller I had built to play
the game in the console showed that.
I also wired the animations, but I had a hard time getting them to sync
up properly and not look disjointed.
I had to resort to creating a sleep function that actually behaved like
a traditional sleep function.
Lastly, I had to add an overlay during animations to prevent unwanted
input at the wrong times.
The game is functionally complete including the animations.
I will add one last thing to be requirement complete, and that is the
ability to change players names and display scores.